Fix string quoting for selector problems
E.g. if the name of a hyperref contains a blank, you have to quote that
name. If you didn't use apostrophes but quotation marks it would break
the code in Ghost.click(...) and related method.

By using `repr(str)` you won't have this problem anymore, as Python
quotes the string properly, i.e. in a way that makes JavaScript happy.