Log download progress of responses
Some responses take a while to download so have some logs to see what is
going on. This code should probably be enhanced to skip small downloads
and or start emitting logs if downloads takes more than a pre-defined
amount of time but for now it is more helpful as is to help debug
network problems in the current code.