Make QNetworkReply callbacks more resilient
So they can log when reply and/or manager is destroyed already. While
NetworkAccessManager.__del__ considerably reduces the occurence of
this problem, it is still pops up once in a while.

Ultimately, implementing a custom QNetworkReply class should probably
solve this as both python and C++ code would rely on the same object to
exist and execute required code.